requirements.txt
discord
requests
json
python-dotenv
ssl
certifi